move clubs to own app
