+++ README.md
@@ -58,9 +58,9 @@ App is reachable under [127.0.0.1:8000](http://127.0.0.1:8000)
 ## Main Management Commands
 
 - `import_associations`: Associations
+- `import_clubs`: Clubs
 - `import_districts`: Districts
 - `import_leagues`: Seasons, Leagues, Teams
-- `import_clubs`: Clubs
 - `import_games`: Games, Sport Halls
 - `import_reports`: Players, Scores
