+++ src/hbscorez/urls.py
@@ -9,6 +9,7 @@ from api import urls as api_urls
 from associations import urls as association_urls
 from associations.models import Association
 from base import urls as base_urls
+from clubs.models import Club
 from districts import urls as district_urls
 from districts.models import District
 from games import urls as game_urls
@@ -19,15 +20,15 @@ from players import urls as player_urls
 from players.models import Player
 from sports_halls import urls as sports_hall_urls
 from teams import urls as team_urls
-from teams.models import Club, Team
+from teams.models import Team
 from users import urls as user_urls
 from users.forms import CaptchaRegistrationForm
 from users.views import CaptchaRegistrationView
 
 ASSOCIATIONS = {'queryset': Association.objects.get_queryset().order_by('pk')}
+CLUBS = {'queryset': Club.objects.get_queryset().order_by('pk')}
 DISTRICTS = {'queryset': District.objects.get_queryset().order_by('pk')}
 LEAGUES = {'queryset': League.objects.get_queryset().order_by('pk')}
-CLUBS = {'queryset': Club.objects.get_queryset().order_by('pk')}
 TEAMS = {'queryset': Team.objects.get_queryset().order_by('pk')}
 GAMES = {'queryset': Game.objects.get_queryset().order_by('pk')}
 PLAYERS = {'queryset': Player.objects.get_queryset().order_by('pk')}
@@ -41,9 +42,9 @@ urlpatterns = [
     path('benutzer/', include(user_urls)),
     path('sitemap.xml', sitemap, {'sitemaps': {
         'associations': GenericSitemap(ASSOCIATIONS),
+        'clubs': GenericSitemap(CLUBS),
         'districts': GenericSitemap(DISTRICTS),
         'leagues': GenericSitemap(LEAGUES),
-        'clubs': GenericSitemap(CLUBS),
         'teams': GenericSitemap(TEAMS),
         'players': GenericSitemap(PLAYERS),
     }}, name='django.contrib.sitemaps.views.sitemap'),
